Improves efficiency of hasBetterAudio by making it seqential rather than parallel
Before, this command would run, it would go pretty fast, but it would also make any other networking tasks not work properly.

By changing it to be sequential, it lets the CPU take a break; other networking traffic can proceed to flow, and it's less likely to cause problems when you have a large number of threads available (32+).